Python/fab/requirements.txt
flask-appbuilder
flask-migrate
pymysql
cryptography==3.4.7
python-dotenv
